requirements.txt
Flask==1.0.2
pyserial==3.4
python-dateutil==2.8.0